Kinde Auth

+++ app/actions.ts
@@ -3,6 +3,11 @@
 import Word from "@/models/Word";
 import dbConnect from "@/lib/mongodb";
 import { revalidatePath } from "next/cache";
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
+import { checkPermissions } from "./api/auth/auth-utils";
+import { KindeAccessToken } from "@kinde-oss/kinde-auth-nextjs/types";
+import { MOONCAKE_PERMISSIONS } from "./api/auth/app-permissions";
+import { SentenceType } from "@/types/SajeonTypes";
 
 /**
  *
@@ -23,58 +28,80 @@ export async function formAction(formData: FormData) {
  * that will update the database and respective word.
  */
 export async function updateDatabase(formData: FormData) {
-  /**
-   *
-   * @returns an array of sentences to replace the DB entry.
-   */
-  const formatSentenceObject = (): { kr: string; en: string }[] => {
-    let updatedSentences: { kr: string; en: string }[] = [];
-    for (const [key, val] of formData.entries()) {
-      // target sentences only
-      if (key.includes("kr") || key.includes("en")) {
-        const _strSentenceIndex: string | undefined = key.split("-").pop();
-        const sentenceIndex: number = Number(_strSentenceIndex) - 1;
+  const { isAuthenticated, getAccessToken } = getKindeServerSession();
+  const token: KindeAccessToken | undefined = await getAccessToken();
+  const hasPermissions = checkPermissions(token, [MOONCAKE_PERMISSIONS.edit]);
 
-        // if that entry doesn't exist, create item in array before we push values
-        if (!updatedSentences[sentenceIndex]) {
-          updatedSentences.push({ kr: "", en: "" });
-        }
+  if (!(await isAuthenticated()) || !hasPermissions) {
+    return {
+      error: "Unauthorized",
+      statusCode: 401,
+    };
+  }
 
-        if (key.includes("kr")) {
-          updatedSentences[sentenceIndex].kr = String(val);
-        } else {
-          updatedSentences[sentenceIndex].en = String(val);
+  // permissions granted, authenticaed user with permissions
+  else {
+    /**
+     *
+     * @returns an array of sentences to replace the DB entry.
+     */
+    const formatSentenceObject = (): SentenceType[] => {
+      let updatedSentences: SentenceType[] = [];
+      for (const [key, val] of formData.entries()) {
+        // target sentences only
+        if (key.includes("kr") || key.includes("en")) {
+          const _strSentenceIndex: string | undefined = key.split("-").pop();
+          const sentenceIndex: number = Number(_strSentenceIndex) - 1;
+
+          // if that entry doesn't exist, create item in array before we push values
+          if (!updatedSentences[sentenceIndex]) {
+            updatedSentences.push({ kr: "", en: "" });
+          }
+
+          if (key.includes("kr")) {
+            updatedSentences[sentenceIndex].kr = String(val);
+          } else {
+            updatedSentences[sentenceIndex].en = String(val);
+          }
         }
       }
+
+      return updatedSentences;
+    };
+
+    /**
+     *  TODO: form validation and checks on server once we get
+     *        mooncakes fixed up. Server & Client should both use validation.
+     *        perhaps zod or something similar.
+     *  DictionaryEntryType
+     *
+     *  */
+
+    const updatedData = {
+      _id: formData.get("_word-id"),
+      word: formData.get("word"),
+      romaja: formData.get("romaja"),
+      hanja: formData.get("hanja"),
+      definitions: formData.getAll("definition"),
+      // explanation: formData.get('romaja'),
+      pos: formData.get("pos"),
+      sentences: formatSentenceObject(),
+    };
+
+    await dbConnect();
+    try {
+      // Find the Word record by ID and update it in the DB
+      await Word.findByIdAndUpdate(updatedData._id, updatedData);
+      console.log(`Succesfully updated record ${updatedData}`);
+      revalidatePath("/");
+    } catch (error) {
+      return new Response(JSON.stringify({ message: (error as any).message }), {
+        status: 500,
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
     }
-
-    return updatedSentences;
-  };
-
-  const updatedData = {
-    _id: formData.get("_word-id"),
-    word: formData.get("word"),
-    romaja: formData.get("romaja"),
-    hanja: formData.get("hanja"),
-    definitions: formData.getAll("definition"),
-    // explanation: formData.get('romaja'),
-    pos: formData.get("pos"),
-    sentences: formatSentenceObject(),
-  };
-
-  await dbConnect();
-  try {
-    // Find the Word record by ID and update it in the DB
-    await Word.findByIdAndUpdate(updatedData._id, updatedData);
-    console.log(`Succesfully updated record ${updatedData}`);
-    revalidatePath("/");
-  } catch (error) {
-    return new Response(JSON.stringify({ message: (error as any).message }), {
-      status: 500,
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
   }
 }
 
@@ -84,23 +111,45 @@ export async function updateDatabase(formData: FormData) {
  * @description This action will take a word ID and remove it from the database.
  */
 export async function deleteWord(formData: FormData) {
-  const _id = formData.get("_word-id");
+  const { isAuthenticated, getAccessToken } = getKindeServerSession();
+  const token = await getAccessToken();
+  const hasPermissions = checkPermissions(token, [MOONCAKE_PERMISSIONS.edit]);
 
-  if (_id) {
-    await dbConnect();
+  if (!(await isAuthenticated()) || !hasPermissions) {
+    return {
+      error: "Unauthorized",
+      statusCode: 401,
+    };
+  } else {
+    /**
+     *  TODO: form validation and checks on server once we get
+     *        mooncakes fixed up. Server & Client should both use validation.
+     *        perhaps zod or something similar.
+     *  DictionaryEntryType
+     *
+     *  */
 
-    try {
-      // Find the Word record by ID and update it in the DB
-      await Word.findByIdAndDelete(_id);
-      console.log(`Succesfully deleted record ID: ${_id}`);
-      revalidatePath("/");
-    } catch (error) {
-      return new Response(JSON.stringify({ message: (error as any).message }), {
-        status: 500,
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
+    const _id = formData.get("_word-id");
+
+    if (_id) {
+      await dbConnect();
+
+      try {
+        // Find the Word record by ID and update it in the DB
+        await Word.findByIdAndDelete(_id);
+        console.log(`Succesfully deleted record ID: ${_id}`);
+        revalidatePath("/");
+      } catch (error) {
+        return new Response(
+          JSON.stringify({ message: (error as any).message }),
+          {
+            status: 500,
+            headers: {
+              "Content-Type": "application/json",
+            },
+          },
+        );
+      }
     }
   }
 }
